[+] generic self baff ability [+] second wind ability [~] spike and spike bushes abilities now use new generic script
