+++ Assets/Content/Prefabs/Abilities/Active Abilities/Second Wind Ability.prefab
@@ -11,10 +11,10 @@ PrefabInstance:
     - target: {fileID: 5863459408028885802, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
       propertyPath: material
       value: 
-      objectReference: {fileID: 2100000, guid: f124592f99263ec448bff478012110a8, type: 2}
+      objectReference: {fileID: 2100000, guid: dae6e9382839e1e428ea6c4afb64dc90, type: 2}
     - target: {fileID: 5863459408028885802, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
       propertyPath: abilityName
-      value: "\u0422\u0435\u0440\u043D\u043E\u0432\u044B\u0435 \u0437\u0430\u0440\u043E\u0441\u043B\u0438"
+      value: "\u0412\u0442\u043E\u0440\u043E\u0435 \u0434\u044B\u0445\u0430\u043D\u0438\u0435"
       objectReference: {fileID: 0}
     - target: {fileID: 5863459408028885804, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
       propertyPath: m_RootOrder
@@ -62,7 +62,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5863459408028885805, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
       propertyPath: m_Name
-      value: Spike Bushes Ability
+      value: Second Wind Ability
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
@@ -70,7 +70,7 @@ PrefabInstance:
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 5863459408028885805, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 5774373266950952726}
+      addedObject: {fileID: 5785306727670875927}
   m_SourcePrefab: {fileID: 100100000, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
 --- !u!114 &8592677024977933672 stripped
 MonoBehaviour:
@@ -88,7 +88,7 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 5863459408028885805, guid: 42a4728cc87fcae4793ebb6eeaaf76be, type: 3}
   m_PrefabInstance: {fileID: 2765319565851473474}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &5774373266950952726
+--- !u!114 &5785306727670875927
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -109,16 +109,12 @@ MonoBehaviour:
   cooldownAfterStart: 0
   cooldownCounter: 0
   isCooldownedAfterStart: 0
-  abilityAnimatorTrigger: 0
+  abilityAnimatorTrigger: 3
   weaponIndex: -1
   timeBeforeImpact: 0.15
   fullAnimationTime: 0.8
   delayBeforeImpactSound: 0
   impactSoundSource: {fileID: 0}
-  temporaryBlockValue: 1
-  permanentBlockValue: 1
-  Buffs:
-  - BuffType: 40
-    BuffValue: 1
-    TimeToEndBuff: 1
-    BuffIsConstantly: 0
+  temporaryBlockValue: 0
+  permanentBlockValue: 0
+  Buffs: []
