+++ Assets/Content/Prefabs/Abilities/Active Abilities/Second Wind Ability.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f9266cc6b7d9c724ca908b413bb1e633
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
