+++ Assets/Content/Prefabs/UI/Ability Button/Materials/Second Wind Button.mat
@@ -20,7 +20,7 @@ Material:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Spike Bushes Button
+  m_Name: Second Wind Button
   m_Shader: {fileID: -6465566751694194690, guid: 8396ab3bf3260894a98de82e1227ac61, type: 3}
   m_Parent: {fileID: 0}
   m_ModifiedSerializedProperties: 0
@@ -45,7 +45,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: da08f1a03c0d0454f8737ccd90016525, type: 3}
+        m_Texture: {fileID: 2800000, guid: a4317138c0230334b9eff5d99fa01e3f, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _Underline_Highlighing:
