+++ Assets/Content/Prefabs/UI/Ability Button/Materials/Second Wind Button.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dae6e9382839e1e428ea6c4afb64dc90
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
