--- Assets/Scripts/Abilities/Ability Instances/SpikeBushesAbility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27fee5c97d872a048baf7c592f52bae4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
